move installation instructions to tutorial package
The tutorial needs to be cloned and run locally so the local installation
instructions seem suitable for it.

Now, for the global EDIT installation, I am not sure that would be the best way
to go (e.g. compared to installing from Pypi), meaning that the main README
will likely need to be rewritten once EDIT is open-sourced. For now, I am
removing this README as it is a bit of a duplication with the tutorial README,
but happy to put it back if needed.