pin dask version and provide locked environment definition file
It seems that recent version of Dask 2024.12.0 slows down a lot the CNN
training notebook. This is not the case for an earlier version 2024.9.1. For
now, we pin the version of Dask in pyproject.toml.

Also, this commit provides a fully resolved environment.lock.yml file to,
hopefully, provide better reproducibility.