remove main package, move its definition to root folder
The `pyearthtools_main" package is an empty packaged used to aggregate the
other ones, installing all of them together with a common minimum version.

Moving the pyproject.toml at the root allow it to be also used for other
purposes, for example to define dependencies to build the docs.

I have not moved the only piece of code that was in __init__.py, which was used
to list versions. I have not found this function used anywhere and it can be
replace with a `pip list`or equivalent command.