remove a level of folder nesting, make ERA5 valid variables a list
The valid variables file "pyearthtools.data.download.arco.variables.ERA5.valid"
is only used in this context (i.e. validate ARCOERA5.__init__ inputs) and can
be replaced by a list variable. Using this, I regrouped it in the same folder
as the valid ERA5 valid pressure levels - which were alread a variable and not
a .valid file - and removed a level of folder nesting which wasn't needed
anymore.