replace global object with loading function
Global variables introduce an hidden side channel that might not be desirable.
I understand here that is was use to save loading time (it took 1m44s to open
the full dataset for me). This could be replaced by using functools.cache, that
properly reload if the kwargs have changed, or by fixing the loading time by
dropping the variables not needed.