precompute reverse name mapping and create variable list from them
ERA_NAME_CHANGE contains all variables names, long (keys) and short (values),
so it can readily be used to validated the input of ARCOERA5, no need to
duplicate that list.

In addition, the reverse mapping can be precomputed and use as is, instead of
introducing a function (with a nested function) for this.