reorganise installation page
I have moved the virtual environmnt explanations at the end of the page, as
they are more general knowledge that we can reference in multiple places for
users that want to know more about it. This way, installation instructions are
more readily accesible (less scrolling).

I also added the virtual/conda environment tabs in the developer section.

And I have updated the intro to mention the NCI user/dev case.