download license and display message about it
I used "print" as this message is intended for the user and we don't want the
logging system to silence it (actively or because a handler has not been
attached to the logging system). for now, I am using a constructor parameter
to disable the message, this avoids setting up a configuration scheme.